% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afrb-dir.R
\name{afrb_dir}
\alias{afrb_dir}
\title{Set the Afrobarometer data directory}
\usage{
afrb_dir(path = tempdir())
}
\arguments{
\item{path}{a string containing the file path to set as the Afrobarometer data directory
\code{getOptions("afrobarometer.data")}). Run \code{afrb_dir()} without any parameters
to use \code{path = base::tempdir()}.}
}
\value{
Sets the Afrobarometer data directory and invisibly returns the path
to the directory as a string.
}
\description{
Initialize the Afrobarometer data directory for the session.
}
\details{
Sets the options the Afrobarometer data directory: \code{afrobarometer.data}.
This directory is where raw and merged data is cached across sessions. \code{afrb_dir(path)}
creates the following subdirectories:
\itemize{
\item{codebooks}{Codebooks (PDF) for each round are downloaded and stored in \code{path/codebooks}.}
\item{locations}{Afrobarometer restricted access georeferenced data (CSV) should be stored in
      \code{path/locations}. The user is expected to manually place each CSV file of georeferenced
      data in \code{locations} like \code{path/locations/Location_R*.csv}.}
\item{questionnaires}{Afrobarometer survey questionnaires (SAS) are downloaded and stored in \code{path/questionnaires}.}
}
}
\examples{

library(afrobarometer)
afrb_dir("~/foo")

}
