% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afrb-round.R
\name{afrb_round}
\alias{afrb_round}
\alias{afrb_list_rounds}
\title{Load data for an Afrobarometer Survey round as a local tibble}
\usage{
afrb_round(round = NULL)

afrb_list_rounds()
}
\arguments{
\item{round}{An integer of the round to select.  \code{round \%in\% 1:6} must be true}
}
\value{
\code{afrb_round(x)} returns a \code{data.frame} for Round ]code{x}.  \code{afrb_list_rounds}
returns a character vector of file paths for built datasets from \code{\link{afrb_build}},
if any exist.
}
\description{
After the local database has been built (\code{\link{afrb_build}}), \code{afrb_round}
reads data for the specified round into the workspace as a \code{data.frame}.
}
\details{
\code{afrb_round} loads built data for round as a \code{data.frame}.  Assumes
\code{\link{afrb_build}} has been run.  \code{afrb_list_rounds} lists built datasets in the
Afrobarometer data directory, if any exist.
}
\examples{
\dontrun{
library(afrobarometer)
afrb_dir("~/foo")
afrb_build(rounds = 3, overwrite = TRUE)
r3 <- afrb_round(round = 3)
}

}
