% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{afrb_build}
\alias{afrb_build}
\title{Build a local copy of the Afrobarometer database}
\usage{
afrb_build(rounds = 1:6, overwrite_db = FALSE)
}
\arguments{
\item{rounds}{A sequence of integers indicating which rounds to build.  As
of February 2017, there are six rounds.  Default is \code{seq(1: 6)}.}

\item{overwrite_db}{A boolean value indicating whether or not the SQLite
database in the local Afrobarometer data directory should be overwritten 
when \code{afrb_build} is run.  Default is \code{FALSE}.}
}
\value{
Builds the database in \code{options("afrobarometer.data")},
downloading files as needed into subdirectories and merging the resulting
data in the SQLite database file \code{options("afrobarometer.sqlite")}.
}
\description{
Automates the downloading and merging of the Afrobarometer data.  Fetches
any missing, publicly available files, creates a master SQLite database in
\code{options(afrobarometer.data)}, and merges available spatial data with
the questionnaires for each round.
}
\details{
Meant to be run after \code{afrb_dir()} and before
\code{afrb_round}.
}
\examples{
\dontrun{
library(afrobarometer)
afrb_dir("~/foo")
# download first 3 rounds but do not overwrite the existing database file.
# This will fail if the database file has tables.
afrobarometer::afrb_build(1:3)
# download all rounds and overwrite the existing database file.
afrobarometer::afrb_build(overwrite_db = TRUE)
}

}
\seealso{
\code{\link{afrb_dir}}
}
