% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afrb-build.R
\name{afrb_build}
\alias{afrb_build}
\title{Build a local copy of the Afrobarometer database}
\usage{
afrb_build(rounds = 1:6, overwrite = FALSE)
}
\arguments{
\item{rounds}{A sequence of integers indicating which rounds to build.  As
of February 2017, there are six rounds.  Default is value \code{1:6}.}

\item{overwrite}{A boolean value indicating whether or not the SQLite
database in the local Afrobarometer data directory should be overwritten
when \code{afrb_build} is run.  Default is \code{FALSE}.}
}
\value{
Builds the database in \code{options("afrobarometer.data")},
downloading files as needed into sub-directories and merging the resulting
data \code{file.path(options("afrobarometer.data"), "build")}.
}
\description{
Automates the downloading and merging of the Afrobarometer data.  Fetches
any missing, publicly available files, creates a master MonetDB database in
\code{options(afrobarometer.data)}, and merges available spatial data with
the questionnaires for each round.
}
\details{
May only run \code{afrb_round} after \code{afrb_dir} and before
\code{afrb_round}.
}
\examples{
\dontrun{
library(afrobarometer)
afrb_dir(path = "~/foo")
# download first 3 rounds but do not overwrite the existing database.
# This will fail if the database is already populated with merged tables.
afrb_build(rounds = 3:4)
# download all rounds and overwrite the existing database.
afrb_build(rounds = 3:4, overwrite = TRUE)
r3 <- afrb_round(round = 3)
}

}
\seealso{
\code{\link{afrb_dir}}
}
