% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{afrb_dir}
\alias{afrb_dir}
\alias{afrb_dir}
\title{Set the Afrobarometer data directory}
\usage{
afrb_dir(path = tempdir())
}
\arguments{
\item{path}{Character scalar containing the file path to set. Run 
\code{afrb_dir()} without any paramters to use \code{base::tempdir()}.}
}
\value{
Sets the afrobarometer data directory and database file invisibly.
}
\description{
Customize where the data is cached.  Must be run explicitly by the user.
}
\details{
Sets the options \code{afrobarometer.data} and
\code{afrobarometer.sqlite}, the paths of the local Afrobarometer data cache
and SQLite database respectively.  The questionnaire data will be
automatically downloaded into the subdirectory \code{questionnaires}.  The
user is expected to manually add spatial data CSV files in the format
\code{locations/Location_Rk.csv} where \code{locations} is a subdirectory of
the Afrobrometer data folder, \code{afrobarometer.data}.  \code{afrb_dir()}
automatically creates \code{locations}, so the user simply has to place the
spatial data CSVs in the \code{locations} subdirectory after running
\code{afrb_dir()}.
}
\examples{

library(afrobarometer)
afrb_dir("~/foo")

}
