% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.R
\name{afrb_round}
\alias{afrb_round}
\title{Load data for an Afrobarometer Survey round as a local tibble}
\usage{
afrb_round(round)
}
\arguments{
\item{round}{An integer of the round to select.  As of February 2017, there
is questionnaire data available for Rounds 1 through 6.}
}
\value{
tbl A local tibble with data for the round.  If spatial data is
available for the round, then the objet returned by
\code{afrobarometer::afrb_round(x)} will be merged spatial and questionnaire
data for round \code{x}.
}
\description{
After the local database has been build, this function reads data for the
specified round into R as a local tibble.
}
\examples{
\dontrun{
library(afrobarometer)
afrb_dir("~/foo")
afrb_build(overwrite = TRUE)
r3 <- afrb_round(3)
}
}
